Use FlatList for some address lists
